add class digraph

+++ src/basic/graph.h
@@ -12,13 +12,13 @@ namespace PlanarGraphColoring {
 class Graph {
 protected:
   size_t n_;/// number of vertices indexed from 0 to n-1
-  std::unordered_map<int, std::vector<int>> neighbors_;
+  std::unordered_map<size_t, std::vector<size_t>> neighbors_;
 
 public:
   Graph() = default;
-  Graph(const size_t n, const std::vector<std::vector<int>>& edges);
+  Graph(const size_t n, const std::vector<std::vector<size_t>>& edges);
   virtual void show() const;
-  std::vector<int> getNeighbors(const int vertex) const;
+  std::vector<size_t> getNeighbors(const size_t vertex) const;
   inline size_t size() const {
     return n_;
   }
